Add A/B test variants for SecretLinksTableRow
Introduce variant loader pattern for testing different row designs:
- timeline: Original design with badge status, icon actions, connector
- console: Monospace design with tree metadata, text buttons

Loader uses async imports for code splitting. Default set to console
variant for initial testing.
